big update
it actually works now.
added support for disabling the microphone and entering words with the keyboard itself.
refactored the code in many ways, almost a whole rewrite
created a simple replacements language for editing the replacements file in Notepad++.
added support for separating a statement into multiple statements with the "then" keyword